some changes

+++ backend/controllers/category.controller.js
@@ -4,23 +4,16 @@ import { ApiErrors } from "../utils/ApiErrors.js";
 import { ApiResponce } from "../utils/ApiResponce.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
 
-const createCategory = asyncHandler(async (req, res) => {
-    const { categoryname } = req.body;
-    // if (!categoryname)
-    //     throw new ApiErrors(400, "category name  cannot be empty");
-
+const createNewCategory = asyncHandler(async (req, res, next) => {
     const errors = validationResult(req);
-    console.log("validation  result", errors);
-
     if (!errors.isEmpty()) {
-        return res.status(400).json({ success: false, error: errors.array() });
-        // throw new ApiErrors(400, "validation error", errors);
+        return res
+            .status(400)
+            .json(new ApiErrors(400, "bad request", errors.array()));
     }
-
-    const catExist = await Category.findOne({ categoryname: categoryname });
-    if (catExist) throw new ApiErrors(409, "This category already exist");
-
-    const newCategory = await Category.create({ categoryname });
+    const newCategory = await Category.create({
+        categoryname: req.body.categoryname,
+    });
 
     if (!newCategory)
         throw new ApiErrors(401, "failed to create a new categiry");
@@ -30,33 +23,6 @@ const createCategory = asyncHandler(async (req, res) => {
     );
 });
 
-const createNewCategory = async (req, res, next) => {
-    const errors = validationResult(req);
-    if (!errors.isEmpty()) {
-        return res.status(400).json({
-            success: false,
-            errors: errors.array(),
-        });
-    }
-    try {
-        const newCategory = await Category.create({
-            categoryname: req.body.categoryname,
-        });
-
-        if (!newCategory)
-            throw new ApiErrors(401, "failed to create a new categiry");
-
-        res.status(200).json(
-            new ApiResponce(201, newCategory, "category created sucessfully")
-        );
-    } catch (error) {
-        res.status(400).json({
-            success: false,
-            errors: error.message,
-        });
-    }
-};
-
 const allCategories = asyncHandler(async (req, res) => {
     const categories = await Category.find();
     if (!categories) throw new ApiErrors(404, "No Categories Found!");
@@ -64,6 +30,12 @@ const allCategories = asyncHandler(async (req, res) => {
 });
 
 const getCategory = asyncHandler(async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res
+            .status(400)
+            .json(new ApiErrors(400, "bad request", errors.array()));
+    }
     const category = await Category.findById(req.params.id);
     if (!category) throw new ApiErrors(404, "no category with this Id");
 
@@ -71,12 +43,19 @@ const getCategory = asyncHandler(async (req, res) => {
 });
 
 const updateCategory = asyncHandler(async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res
+            .status(400)
+            .json(new ApiErrors(400, "Bad request", errors.array()));
+    }
     const { categoryname } = req.body;
-    if (!categoryname)
-        throw new ApiErrors(400, "Please provide the field to update");
-    const catExists = await Category.findOne({ categoryname });
-    if (catExists)
-        throw new ApiErrors(409, "The provided category is already in use");
+
+    // if (!categoryname)
+    //     throw new ApiErrors(400, "Please provide the field to update");
+    // const catExists = await Category.findOne({ categoryname });
+    // if (catExists)
+    //     throw new ApiErrors(409, "The provided category is already in use");
 
     const updatedCategory = await Category.findByIdAndUpdate(
         req.params.id,
@@ -91,6 +70,12 @@ const updateCategory = asyncHandler(async (req, res) => {
 });
 
 const deleteCategory = asyncHandler(async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res
+            .status(400)
+            .json(new ApiErrors(400, "bad request", errors.array()));
+    }
     const delCategory = await Category.findByIdAndDelete(req.params.id);
     if (!delCategory) throw new ApiErrors(404, "No such category found.");
 
@@ -100,7 +85,6 @@ const deleteCategory = asyncHandler(async (req, res) => {
 });
 
 export {
-    createCategory,
     allCategories,
     getCategory,
     updateCategory,
